control.c: Minor fix to malloc correct data structure size.
Memory is allocated for mrm_host_list, which is a pointer to an array
of pointers to linked_item.  Currently add_linked_item mallocs a full
linked_item structure for each array entry instead of the required
pointer to linked_item.  linked_item is larger than *linked_item so
no harm done, but it wastes a bit of memory.