Convert hostgroup selection structure to store pointers in hash table.
Currently the actual integer is stored in the hash table values.  This
causes two problems.  The first is that on 64-bit systems you get data
size mismatch warnings.  The second and more serious problem is that
the hash_find functions in hash.c return a 0 for non-existant hash entries.
However, 0 is also a valid selection integer.  This patch converts the
selection entries to be malloc()d integers which addresses both issues.