daemon.c: Fix minor memory leak
In import_objects_and_status asprintf() is used to allocate and create
a string.  In some situations it then calls asprintf again to essentially
concatenate a new string on the previous string.  However, it stores the new
allocated string over the pointer to the previous string, causing the memory
for that string to be unable to be free()d.  This patch corrects that.