sql.c: Convert sql_error allocated string to string constant.
Currently in sql_error() if there is no db.conn it returns a pointer to an
allocated string.  This pointer is not free()d later, nor is safe to free()
since it could also be a pointer to a string returned from the library
function dbi_conn_error.  The documentation for dbi_conn_error states that
the string should not be free()d by the application.  This patch converts
the error message to a string constant, which will prevent any memory leaks.