Add configuration option "ipc_binlog" for configurable binlog path.
This adds a new configuration option "ipc_binlog" allows the
specification of the path where binlogs are written to disk.
This is configurable in both the module and daemon configuration
groups.  It also checks to make sure that the path is usable by
performing a test open on the path to help ensure configuration
problems are caught at start-time instead of when it actually
tries to write to the binlogs.

Signed-off-by: Andreas Ericsson <ae@op5.se>