Fix poller node automatic (de)activate ipc message sending.
There is support for the daemon to send IPC control messages to
the module to dynamically enable/disable checks on the NOC for
hostgroups based on whether the poller associated with that
hostgroup is connected or not.

Currently, the code doesn't trigger the transition at key
critical connection events with the poller.  Also, if Nagios
was restarted without restarting the daemon there is no
facility to remember the last state sent and to trigger a
re-notification of the module to (de)activate the checks
properly.

This patch refactors the support a bit to add tracking of
the last state sent, trigger the previously unused node
action function to set the state at all connection events,
and re-notifies the module via IPC of the last state for
the hostgroups after each IPC re-connection.

Signed-off-by: Sean Millichamp <sean@bruenor.org>