object importer: Cache contact access rights at import-time
This doesn't take very long, but simplifies the ninja queries
immensely. Only the contacts that are assigned to hosts or services
are cached in the newly created contact_access table. Contacts that
by some other means are allowed to see absolutely everything aren't
handled by this cache.

This fixes issue #2883.

Signed-off-by: Andreas Ericsson <ae@op5.se>