sql.c: Don't shutdown and re-initialize dbi upon reconnect
Doing so causes libdbi to leak memory, and we don't really need to
do it once we've loaded the drivers etc anyway.

This means we'll never do dbi_shutdown() at all, and will "leak"
memory when closing down. However, the kernel will immediately
release it back to the pool of free memory anyway, so the leak has
no real-world impact (in fact, fixing it would cause a slight
performance degradation).

Signed-off-by: Andreas Ericsson <ae@op5.se>