binlog: Change size_t to uint
It doesn't make sense to not utilize the architectures default fastest
type in favour of one that's often smaller and may be slower to access.

We also rename max_size and size to max_mem_size and mem_size,
respectively, in the binlog structure, since this better reflects
what it is they measure.

Signed-off-by: Andreas Ericsson <ae@op5.se>