sql: Be more careful when quoting strings
Previously we used to disregard the return value of dbi when asking it
to quote and escape strings. Since dbi_conn_quote_string() can fail,
we need to catch such errors and make sure we log and dodge them.

With this patch, the string we failed to quote will simply be left
empty, and the source and destination pointers will be logged, along
with the string we failed to quote (in a separate message).

Signed-off-by: Andreas Ericsson <ae@op5.se>