db_updater: Plug memory leaks from inserting notifications
sql_quoted() strings are always duplicated, but we failed to free()
the memory they used earlier for the case when we were inserting
contacts. This is likely the cause of crashes on customer systems
with long-running merlin daemons, since our beta-systems stupidly
enough aren't configured to even try to send notifications.

Signed-off-by: Andreas Ericsson <ae@op5.se>