mtest: Update host and service tests to run on all available objects
This lets us make sure the blockify()/deblockify() function pair
doesn't fsck up on a particular set of strings found in the wild.

We also introduce verify_count(), which is added to check the rowcount
returned by a certain query. This is useful to verify that all events
are being duly processed and the database is updated the way we want
it to be.

Signed-off-by: Andreas Ericsson <ae@op5.se>