mtest: Add zzz() macro
This provides a short period of sleep which is roughly twice the
longest measured amount of time required for an event to be created,
sent, received, parsed and inserted into the database.

It's useful to use before verifying that an event caused the database
to be updated the way we want it to be.

Signed-off-by: Andreas Ericsson <ae@op5.se>