mtest: Optimize count_rows()
It's no use fetching all resulting rows when we can just as well use
the libdbi function dbi_result_get_numrows() to check how many rows
were returned. This reduces the runtime of the tests to a much more
manageable level and doesn't cause it to eat CPU like mad.

Signed-off-by: Andreas Ericsson <ae@op5.se>