module: Add support for being run through mtest
This adds basic support for running the module initialization process
through mtest. If certain magic parameters are passed, we ignore
changing directory to somewhere where we know we can dump core, and
we also ignore reading the configuration, since we know everything
we need will be set by the mtest program.

Signed-off-by: Andreas Ericsson <ae@op5.se>