db_updater: Initialize string vars in handle_downtime() to NULL
This is necessary to prevent a segfault in the merlin daemon in
case we get a NEBCALLBACK_DOWNTIME_DATA event with a type of
NEBTYPE_DOWNTIME_DELETE.

Previously we used to crash horribly at such events. Now they're
handled gracefully. Those events are generated when downtime is
cancelled before it starts or while it's running, so it's not a
very common occurrence.

Signed-off-by: Andreas Ericsson <ae@op5.se>