daemon: Plug a microscopic memory leak
When we received a command to import objects earlier that held not
only a path to the objects.cache file but also a status.log file,
we used to leak the half-finished cmd pointer (roughly 1K).

Even though this leak is very minor and isn't triggered very often,
it's still prudent to make sure we remove it. This patch does just
that. Running through all 9175 tests (46 hosts, 372 services) with
the merlin daemon running under valgrind --leak-check=full proves
that the daemon is now completely free from unreachable blocks.

Signed-off-by: Andreas Ericsson <ae@op5.se>