Squelch valgrind warnings about sending uninitialized data
If a stack-allocated merlin event packet isn't properly initialized
before being sent, valgrind throws a warning. We know the values we're
using are indeed initialized, but making sure valgrind shuts up about
the instances we know about makes it a lot easier to spot the ones we
might introduce in the future that could well be real bugs.

Signed-off-by: Andreas Ericsson <ae@op5.se>