sql: Don't rerun queries on "duplicate index" error
There appears to be a bug in Nagios at the moment, causing
downtime callbacks of type NEBTYPE_DOWNTIME_ADD to be sent
twice. This is rather annoying, since any and all downtime
added will cause us to spam the logs quite a bit and also
to re-connect to the database, with all that it entails.

This patch prevents the reconnecting part and avoids one
rather bulky error message from entering the logs more than
once.

Signed-off-by: Andreas Ericsson <ae@op5.se>