db_updater: Make NEBTYPE_FLAPPING_STOP event remove flapping comments
The flapping data structure has a "comment_id" field which points to
the comment id (doh) associated with this particular flapping state.
In case no additional comment event of the NEBTYPE_COMMENT_DELETE type
is generated, we need to clear up the comments from here. Worst case
scenario, we'll try to remove a comment that doesn't exist, resulting
in a cache miss in the database every time an object stops flapping. Not
exactly a showstopper in terms of lost performance, in other words.

Signed-off-by: Andreas Ericsson <ae@op5.se>