object_importer: Truncate tables scheduled_downtime and comment
Since both downtime and comments are listed in the status.log anyway,
we can safely truncate those tables entirely. This is also a workaround
for issue #3318 on already installed systems and will provide a fix for
this issue in case the merlin daemon for some reason wasn't running when
the downtime stop event occurred.

Signed-off-by: Andreas Ericsson <ae@op5.se>