daemon: Don't prime object states
There's no real need for us to do so since it will be overwritten by
the import program anyway. If the call to prime them should be issued,
it should come after we've run the import program.

This removes one error message about there not being a database
connection from the merlin log when it's just started.

Signed-off-by: Andreas Ericsson <ae@op5.se>