daemon: Make sure all nodes added have a port defined
With this patch we make sure they get the default port if none was
configured in the configuration file. Since default_port is instantiated
with a value of 15551 and can never get a value below 1, we know for
sure that it will always hold a valid port value.

We also change the type of default_port to the more appropriate
unsigned short.

This fixes issue #2630.

Signed-off-by: Andreas Ericsson <ae@op5.se>