protocol.h: Extend the control packets with STALL, RESUME and STOP
STALL will be used to denote "I'm busy, so don't send any more
packets for a while, and stall your process if you can".

RESUME will be used to signal that we're once again ready to
receive events from a previously STALL'ed node.

STOP will be used to signal to a remotely running process that it
should exit immediately.

These new control packets aren't used yet, but will be shortly.

Signed-off-by: Andreas Ericsson <ae@op5.se>