daemon: Stall the module while importing
If we get spammed by events while the import is running, odds are the
(fairly small) socket buffer will get filled up and we'll be asked to
run the import again just as we're about to resume normal operations.

This puts a heavy load on the database, which gets its tables filled,
truncated and indexed repeatedly, and could potentially bring a system
to its knees.

Signed-off-by: Andreas Ericsson <ae@op5.se>