Use an unsigned constant to mean "generic control packet"
Previously we used -1 for this, but since -1 is taken to be an
integer and an int doesn't fit into the 16 bits of data we have,
the comparison was always false on the receiving end, meaning we
logged a warning and ignored the packet. This fixes it so that
we always use a value that fits into the field-type used.

Signed-off-by: Andreas Ericsson <ae@op5.se>