daemon: Revamp how we run the import program
Previously we used system(3) which is quite clumsy and not very good
if you want to catch any underlying errors. This patch makes us rely
on execlp() instead, which provides excellently detailed error messages
in case of failure. As an added bonus, we also gain a very slight
performance improvement.

Signed-off-by: Andreas Ericsson <ae@op5.se>