daemon: Simplify polling loop quite a lot
By getting rid of the start and stop timers and instead inserting
log_event_count() once every minute, we not only log a lot less, but
we also get rid of a silly bug that caused the daemon to insert quite
a lot of junk into the logfiles.

We move the repeated call to its own site, after we've tried to reap
any child-processes and make sure we never bother sending requests to
stall when we really don't need to. This is necessary since it has to
be run once every < 20 seconds, and logs are now only updated once
every minute.

Signed-off-by: Andreas Ericsson <ae@op5.se>