ipc: Fix up binlog handling
If we manage to log an event in the binary ipc backlog, it's not lost
and we shouldn't flag it as an error. We'll still try to connect to the
ipc socket whenever we can, and the binary backlog will be emptied
when we get the chance for it, so nothing is actually lost.

Signed-off-by: Andreas Ericsson <ae@op5.se>