ipc: Fix binlog resend code to work properly
When we're invalidating the binlog, we must return -1 to signal the
error back to the caller from ipc_send_event(). This means events have
gone out of sync and the caller must do whatever it can to prevent the
database from following suit. For the daemon, this isn't really possible.
For the module, this means forcing a re-import.

Signed-off-by: Andreas Ericsson <ae@op5.se>