hooks: Don't try to queue a re-import when stalling
If we do, the daemon might cycle imports more or less indefinitely,
which is obviously a bad thing. With a backlog of 110MiB it'll take
a LOOOONG time before we really need to run a re-import anyway.

While we're at it, we update the error message for when we really need
to run a re-import.

Signed-off-by: Andreas Ericsson <ae@op5.se>