ipc.c: Fix log-spamming when ipc connection is lost
Previously we used to spew quite a lot of repeated entries along the
lines of "Attempting to initialize ipc socket ... " and a couple of
more lines stating why it didn't work in case the daemon has gone to
lunch or died.

With this patch, we add one such set of lines every 30 seconds instead.

Signed-off-by: Andreas Ericsson <ae@op5.se>