ipc: Parse backlog location from configuration
This puts the last finishing touch on fixing issue #2152 by letting
the user specify where to store the backlogs on disk.

This can be set either by a global parameter (to set it to the same
for both daemon and module) or as a section-specific parameter. The
name of the actual file will be automatically generated anyway.

To retain sanity, it's advisable that this option is set to an absolute
path.

Signed-off-by: Andreas Ericsson <ae@op5.se>