binlog api: Never exceed the maximum memory limit
This is actually a lie. We will exceed it if we get added to, but
never with more than sizeof(binlog_entry), which shouldn't be more
than 12 bytes even on 64-bit platforms.

But with this fix we never add stuff to memory if doing so would mean
adding it would exceed the memory limit.

Signed-off-by: Andreas Ericsson <ae@op5.se>