binlog api: Don't force-flush memory entries to disk when opening file
It doesn't hurt to maintain items in memory so long as we don't add to
the ones in memory after we've opened the file and started writing
there. Doing so would cause a logistical nightmare of keeping track of
event id's.

This patch breaks out the memory reading parts to its own function, and
attempts to call that before hitting the file. If there are no events
in the memory binlog, we go look in the file instead.

Signed-off-by: Andreas Ericsson <ae@op5.se>