bltest: Update test program to ensure sequential reading
Previously we only tested add/read, one after the other. With this
patch we test add+add+add+add+add/read+read+read+read etc instead.
This makes it a reliable test to ensure proper sequencing in the
binlog api, and the "partially on disk" tests we have would earlier
never add anything to the disk with the recent changes to the API,
since memory was free()'d and the membased binlog could shrink as
well as grow.

Signed-off-by: Andreas Ericsson <ae@op5.se>