Add ipc_sync_lost() to daemon and module
This function gets called whenever we lose synchronization with
other ipc end. Basically, this is true for when we haven't got
a connection and the backlog somehow fails, which means we have
no way of getting the lost events in sequential order to the
other end of the ipc socket.

The daemon does nothing but logs an error. It is, after all,
pretty much out of options. The module, however, uses this
event to trigger a forced re-import.

We also make sure to only use this new mechanism to trigger a
re-import of object status and data, since it's the only time
we really need to do so in a running merlin instance.

Signed-off-by: Andreas Ericsson <ae@op5.se>