module: Don't use connect()/disconnect() handlers for ipc
Doing so means we'll always force a re-import when the daemon
comes online, even if it, strictly speaking, isn't necessary
to run one (since events in the backlog aren't lost, we may
not need to).

We should probably tune down the backlog a bit though, since
110MiB's worth of data being spewed through the daemon and
into the database will consume quite a lot of cycles in one
go.

Signed-off-by: Andreas Ericsson <ae@op5.se>