Get rid of types.h and move its definitions into protocol.h
This is the first step towards unifying the IO layer so we
always operate on nodes. Since network nodes need a binary
backlog, connect() and disconnect() event handlers and all
that other juicy stuff that the ipc socket needs, we might
as well turn the access methods into a unified api.

The only difference between them is that the module will
(still) never attempt to create a unix domain socket.

Signed-off-by: Andreas Ericsson <ae@op5.se>