binlog api: Introduce binlog_num_entries()
By keeping track of number of entries on disk, we can revamp
binlog_has_entries() so that it returns the number of entries
actually in the log instead of just a binary yes/no thing.

We retain binlog_has_entries() as a macro, pointing to the new
binlog_num_entries().

Signed-off-by: Andreas Ericsson <ae@op5.se>