binlog api: Introduce binlog_unread()
This is fairly straightforward and pretty nifty, since we can now
maintain sequential ordering on messages even if we happen to fail
sending one we read from the backlog. We still don't use the new
functionality, but we will do that later when binlogging is added as
a safety net for all nodes.

Signed-off-by: Andreas Ericsson <ae@op5.se>