Revamp event logging
We keep track of the events per node now, so we mustn't use a static
variable in the logging function to prevent spamming the logs with
event count messages. The timer variable is instead moved to the event
counter structure.

While we're at it, we replace the earlier clumsy tv_delta() function
with one that handles days, hours and minutes as well, since a line
stating that x events have been handled in 2948098 seconds doesn't
really add all that much value.

Signed-off-by: Andreas Ericsson <ae@op5.se>