Unify sending events with the node_* layer
By doing this we get binary backlogging, event-counting and other
goodies for free for all nodes we ever send to, read from or in any
other way communicate with. The ipc vs net api's are still around and
required in order to establish connections and make sure they're still
up, but all the lower-level functionality is now handled completely
by the node layer.

This gives us a much better control over various communication errors,
which leads to less code and therefore, presumably, fewer bugs. It also
means we needn't work so hard to distinguish various errors in the
higher level api's but can instead focus on which nodes we should (try)
to send data to.

Signed-off-by: Andreas Ericsson <ae@op5.se>