Warn for expired variables before throwing a config error
Since we expired ipc_debug_{read,write}, configurations that have
either have been causing both merlind and the module to bomb out in
a fairly hardcore fashion. By checking if the variable is expired
before we make the attempt that might fail hard, we can avoid bombing
out on deprecated variables but instead print them as warnings, as is
proper.

Signed-off-by: Andreas Ericsson <ae@op5.se>