Unify node groking code
This patch lets modules have access to the full set of nodes and
be able to manipulate their state as it sees fit.

Perhaps a shared memory segment with read-only access for the
module would be better for this, but letting the module take care
of it will also get the job done, and with minimal complexity too.

Signed-off-by: Andreas Ericsson <ae@op5.se>