Introduce nodes_by_sel_{name,id}
nodes_by_sel_id() replaces nodelist_by_selection() which used
to reside in net.c, so we remove it in favor of the new one.

The intention, long-term, is to make sure the module can keep
a list of all active and in-active nodes in memory, and get
the list of the nodes (and the state of those nodes) from the
host_hash_table.

Signed-off-by: Andreas Ericsson <ae@op5.se>