shared: s/nsel/num_selections/
It's sensible to use mnemonic names for variables outside anything
but a very small scope. Since this is a file-scoped variable, we're
far better off calling it num_selections, since that's what it
represents.

While at it, we move the static variables to the top of the file,
allowing freestyle code rearranging later.

Signed-off-by: Andreas Ericsson <ae@op5.se>