module: Don't disable or enable checks of hosts and services
We really don't need to do so anymore, since the new logic forces
checks that shouldn't be run to be dropped anyway. This fixes the
inconsistency in the ui where the master server appeared to have a
lot of checks in "disabled" state, while the pollers had them as
active.

We also don't really need the host and service lists, since Nagios
already stores them for us, in the form of hostsmember lists inside
the various hostgroups.

Signed-off-by: Andreas Ericsson <ae@op5.se>