Revamp node statistics tracking somewhat
People are bound to ask "how much data does merlin transfer",
and will most likely want an answer in terms of bytes per
seconds. This patch makes it possible to answer that question
at a later time by adding some additional data tracking stuff
to the node struct.

While we're at it, we fix up so that logged events are tracked
by the statistics-counter and rename and reorganize the various
structs touched by this patch to make their names more mnemonic.

Signed-off-by: Andreas Ericsson <ae@op5.se>