daemon: Send packets to module before updating the database
Since the database updater modifies the packet in order to be able to
use the strings in the event, we must take care to send it to the ipc
socket (or, indeed, anywhere) before running the database update.

We also add a big fat comment where the mrm_db_update() function is
called from, making sure noone else comes along and reorders the calls
later.

Signed-off-by: Andreas Ericsson <ae@op5.se>