daemon: Spend less time deducing if we're connected or not
Since a node only get marked as not connected in the event-sending
layer, we really don't need to bother with checking too closely if
we're connected or not. Previously we used to check once for every
packet we sent, tried to read, or read. An awful waste of cpu cycles,
so we ignore the expensive check if it looks like we're connected
and save ourselves a lot of work.

Signed-off-by: Andreas Ericsson <ae@op5.se>