module: Check for duplicate packets when sending stuff to daemon
Most versions of Nagios has issues with sending duplicate host and
service status data events to the broker API. That imposes an
unnecessary workload on the merlin daemon and the database, so we're
better off avoiding it whenever possible.

We do this by memcmp()'ing each event we send with the one sent before
it. If they're truly identical, event headers and all, we simply ignore
the second (and sometimes third and fourth) event.

Considering the amount of duplicate events Nagios creates, this very
nearly doubles the event throughput capacity of the Merlin system by
cutting the amount of work it has to do in half.

As an extra measure, we also add some fairly advanced debugging code
to track down where the duplicate events are created, and how the
nearly duplicate events differ. This should help us remove the buggy
code in Nagios and increase performance even more.

Signed-off-by: Andreas Ericsson <ae@op5.se>