Log data count sent and received when logging node statistics
This provides the basis for checking how much data is transferred
between various nodes in the network. It would be handy to get a
"bytes per second" in/out count as well, but that can be calculated
by hand so long as we have the raw numbers, which this patch
provides.

Signed-off-by: Andreas Ericsson <ae@op5.se>