daemon: Avoid connection attempts to 'pending' nodes
This patch is a more or less complete rewrite of net_try_connect(),
causing it to take into consideration the 'pending' state.

Previously we used to ignore such states and fire off a new
connect() attempt regardless of whether we had already started
one or not. This was far from optimal and caused a huge storm
of connection attempts and log-messages whenever a node went
offline. Worse though is that the connection attempts all
succeeded one by one, but the two sides could end up in a
sequence where they both discarded a completed connection in
favour of a new (previously pending) one.

This sad state didn't exactly prevent messages from going
through, but caused a lot of unnecessary work for no benefit
at all, so we're better off making sure we never end up in it.

Signed-off-by: Andreas Ericsson <ae@op5.se>