module: Avoid sending near-duplicate comment events
Nagios spawns two comment events each and every time a comment is
added. One COMMENT_ADD and one COMMENT_LOAD. So far, COMMENT_LOAD
seems to be the most accurate of the two, so we ignore COMMENT_ADD
and thereby avoid a lot of warning messages from the module about
duplicate comment_id fields when it runs queries.

Signed-off-by: Andreas Ericsson <ae@op5.se>