Make bt_scan() globally available
It's a tremenduously useful debugging function that we can
utilize from several places. The fact that this patch also
removes its callsites, making the function essentially a
no-op doesn't really diminish its usefulness.

While we're at it, we make its body conditional of us using
glibc >= 2.1, since that's when the backtrace functionality
was introduced.

Catering to other c libraries isn't really interesting, but
it's nice to make sure merlin is at least able to compile
with another c library implementation.

Signed-off-by: Andreas Ericsson <ae@op5.se>