Add the slist api
slist (short for 'sorted list') is a minimal library used for
looking up random objects stashed in a numbered array stored
in memory.

It uses a rather simple binary search algorithm to locate the
sought object.

So far it works by the simple expedient of adding stuff in
random order and then sorting the list before allowing
searches on it.

It's intended to (at a later point) always keep things in a
properly sorted order and be able to disallow non-unique
entries.

Signed-off-by: Andreas Ericsson <ae@op5.se>