module: Fix logmessage when receiving control packets
This patch fixes the annoying, misleading and worrying logmessage
about invalid selection we used to print when we received either a
CTRL_ACTIVE or a CTRL_INACTIVE control packet from a distant node.

Since the merlin daemon rewrites the selection with a node id for us
for such packets, we can identify the node it came from and print that
instead of just "invalid selection".

Signed-off-by: Andreas Ericsson <ae@op5.se>