module: Sort peers and assign id's to them (and us) when they connect
This is fairly straightforward. When the module receives a
CTRL_ACTIVE or CTRL_INACTIVE event and that triggers a difference in
whether or not we consider the node connected, we assign peer id's to
ourself and our peers (assuming we have any peers, ofcourse).

This list will later provide the basis for how we select which of the
checks we should run and which should be handled by our peers.

Signed-off-by: Andreas Ericsson <ae@op5.se>