daemon: Negotiate sockets more frequently
Previously we only negotiated the socket if the already existing one
was in STATE_PENDING, but since we don't know how far along the remote
end is when it comes to processing inbound connection attempts, we must
also negotiate sockets when we're already connected.

This patch takes care of that, and also makes sure we always mark the
socket as being connected when we have an inbound connection attempt
from a valid node.

Signed-off-by: Andreas Ericsson <ae@op5.se>