daemon: Revamp net_negotiate_socket()
We negotiate which socket to use by selecting the connection with the
lowest ip-address. If ip is equal, we use the lowest port number instead.
Since we check our own address on inbound connection attempts and the
remote port for connections initiated by us, this should be a fairly
stable algorithm, since both can't be equal on any sane system.

Signed-off-by: Andreas Ericsson <ae@op5.se>