Introduce node_set_state()
It appears we're failing to run the node state-change hook on
quite a few state-changes. This causes the merlin-module to
not always send the CTRL_ACTIVE code for connecting peers and
pollers, which in turn causes masters and peers to run more
checks than they have to.

This patch should fix that.

Renaming the merlin_node.status field to .state instead is
just to let the compiler find all the instances of manual
assignment to it.

Signed-off-by: Andreas Ericsson <ae@op5.se>