module: Fix peer id calculation for 'last peer'
If we ended up being the last peer in the list, we always assigned
peer_id 0 to ourselves, which was quite bogus.

This patch sets our own peer_id to an impossible value and makes sure
we pick the last slot if we didn't get one in the sorting loop.

Signed-off-by: Andreas Ericsson <ae@op5.se>