daemon: Let the module know which remote node sent the packet
We do this by abusing the 'selection' variable to set the node id.
Since we require that both module and daemon use the same config,
we know for certain that they'll both be referring to the same
node.

Without this patch, the module can't know which node is connecting
and thus can't sort peers properly.

Signed-off-by: Andreas Ericsson <ae@op5.se>