module: Always create ipc connection on-demand
Without this patch, we would sometimes fail to connect at first,
then reschedule the connection attempt and possibly have a message
coming through later that succeeded in creating the connection.

With this patch, we never consciously try to connect to the ipc
layer, but instead let it happen on an ad-hoc basis, forcing the
same mechanism to be used every time. This makes it easier to
debug, and the fact that we're now maintaining less code can only
be considered a Good Thing(tm).

Signed-off-by: Andreas Ericsson <ae@op5.se>