daemon: Don't send control packets to database
The database layer doesn't know what to do with them anyway, so we
shouldn't even make it try.

While we're at it, we add a brief comment noting the need to send
packets to the network before we ship them off to the database handler,
since the database thing always deblockify()'s the event, making it
unusable on the network.

Signed-off-by: Andreas Ericsson <ae@op5.se>