module: Simplify handle_control()
When the daemon forwards packets from other nodes, it will always set
selection to the sending node's id. This lets us simplify the logging
and suchlike quite a lot, since we needn't bother with which selection
something has but can instead focus on the node that sent it.

Assuming noone names a remote node "local Merlin daemon", we'll always
know where a packet originated.

Signed-off-by: Andreas Ericsson <ae@op5.se>