Revampe node and ipc control packet sending systems
Since we'll most likely need to add other control packets
later and the ipc routines will always have to make sure
the connection is up and running before it attempts to send
anything, we're better off adding few very generic functions
instead and then either adding macros or inlined functions
to wrap those few competent functions instead. This is
accomplished with this patch, which introduces ipc_ctrl() and
node_ctrl(), both of which can take an arbitrarily sized
data argument along with a len argument, specifying the size
of the data argument.

The data argument will get copied to the body of the packet,
and the len will ofcourse go into pkt.hdr.len. This patch
already removes more lines of code than it adds, so it's
almost certainly a good move in the long run.

Signed-off-by: Andreas Ericsson <ae@op5.se>