Gain better control of initial ipc variables
Since the ipc node structure is not in the bss but we may still
want to take care to set various variables in it, while nul'ing out
most of it, we introduce a small initialization function, intended
to be run once and once only.

This is the portable way of achieving that result, and also means
we get away with a minimum of maintenance, since memset() uses the
actual size of the structure to zero it all out before we get to
setting anything explicitly.

Signed-off-by: Andreas Ericsson <ae@op5.se>