Convert some macros to inline arguments
Macros that cast any type of pointer variable to a void pointer
are quite evil when it comes to debugging things, as I just
found out to much dismay. Turning them into static inline
functions means no slowdown at all, but the compiler checks to
see that we're passing the correct type of argument before the
inlined function casts it to the void type the ultimate
destination wants.

Signed-off-by: Andreas Ericsson <ae@op5.se>