daemon: Send start-time events to all nodes
Previously we never sent them to master nodes, but that prevented
those master nodes from blocking the checks the pollers were
supposed to run. Since we're talking about 72 bytes here, sent
every time a node connects or disconnects, we really needn't
bother with blocking it for any type of node at all.

If anything, we should in that case block it for pollers, but
they can probably live with reading those 72 bytes as well.

Signed-off-by: Andreas Ericsson <ae@op5.se>