daemon: Handle errors from connect(2) a bit cleverer
Previously we never marked the node as STATE_PENDING even if we
got either EINPROGRESS or EALREADY from connect(). Since that can
only be the result of something severely weird, we can do quite a
bit better than that.

We also handle EISCONN to mark the node as being connected, since
that's exactly what it means.

Signed-off-by: Andreas Ericsson <ae@op5.se>