Introduce node_send_binlog()
Since we'll want to send the binlog as soon as we can rather than
when we get a new event for the targeted node, we need to make it
into its own function. This is quite sensible anyway, since the
functionality to send the binlog is quite large.

While we're at it, we also make sure we never free(temp_pkt) before
accessing it, since doing so will inevitably cause a segmentation
violation when we successfully send packets from the binlog.

Signed-off-by: Andreas Ericsson <ae@op5.se>