io: Handle EAGAIN and EWOULDBLOCK better
Previously we'd return -1 in the case of a 'Resource temporarily
unavailable' error, which could happen quite frequently on slow
network nodes (such as vpn connections or otherwise just long-
distance connections).

With this patch, we wait 100 milliseconds for input on the socket
before trying again when we encounter an event that would block
operations. Basic testing shows the connection is quite stable
now, which is indeed a Good Thing(tm).

Signed-off-by: Andreas Ericsson <ae@op5.se>