Introduce the merlin_nodeinfo struct
This is intended to hold a SHA1 hex of the current running config
along with the last timestamp of that configuration, as well as
the start time for the module, which is used to sort out workloads
between peers.

The information will naturally be used to synchronize configuration
between one (or more) nodes.

Signed-off-by: Andreas Ericsson <ae@op5.se>