daemon: Avoid implicit cast to int when updating program status data
Since the merlin_node.id field is only 16 bits but the sql-query
routine expects an int, the compiler casts it unconditionally and
thus our clever hack to set ipc.id to -1 is suddenly not so clever
anymore, since it would get transformed into 65536 when adding one
to it after the cast to a full 32bit integer.

This patch takes care of it with an explicit check to see if the
update event came from the ipc socket or not, and, if so, setting
the node_id to 0. Otherwise we assign the integer number as per
usual to node->id + 1.

Signed-off-by: Andreas Ericsson <ae@op5.se>