daemon: Mark nodes as unalive when they disconnect or send CTRL_INACTIVE
The whole idea of the program_status table is to let the user know
how the various nagios instances are working at the moment, so we
must take care to also insert 'stop' events when one such node
disconnects. The UI takes care to show nodes that have been
unresponsive for too long as "not running", which may or may not be
the case, but at least gives an indication that something is wrong.

Signed-off-by: Andreas Ericsson <ae@op5.se>