Fix peer id assignment when it occurs multiple times
Previously the peer_id could only ever increase, meaning that if
we received a CTRL_INACTIVE event for one node, as is normal when
one node isn't started yet, we'd run the peer assignment function
twice and all nodes would, eventually, end up with a peer_id the
same as the number of active peers.

This patch corrects it, forcing a hard assignment whenever we run
the sorting list, to make sure that peer_id 0 is always occupied.

Signed-off-by: Andreas Ericsson <ae@op5.se>