io: Be as patient when sending as when reading
In practice, we never seem to wait the maximum 1.5 seconds,
so it's probably safe to do things like this. Hopefully it
will prevent the few disconnects we still see which most
likely happen in case we manage to empty a partially filled
socket before the sending end has time to notice that it's
writable and populate it with more data.

Signed-off-by: Andreas Ericsson <ae@op5.se>