module: Change how we send host and service check results
Since both host and service check results include a pointer to the
object we want to send, and the host and service status updates send
a munged version of that object, we can re-use the host/service status
event code and always send the entire object's pertinent data when we
receive a host or service status update. This also makes it possible to
drop sending host or service status updates later, in case doing so
pushes some bottleneck beyond endurance.
ever gets super-busy.

The daemon still (largely) ignores host and service check result data
events though, but that's a different problem.

Signed-off-by: Andreas Ericsson <ae@op5.se>