module: Handle performance data on check result events
Since we're handling check results events and status update events
the same, this can be done with a single if() clause and a callout
to Nagios' own performance data handling routines, sparing us a
gruesome amount of work we'd be forced to do if we were to implement
this ourselves.

This fixes the module part of issue #2788.

Signed-off-by: Andreas Ericsson <ae@op5.se>